Upgrade govuk-prototype-kit to v13
The app currently errors when starting up with the following message:

```
[...]/dxw-prototype-lib/lib/extensions/extensions.js:75
    throw new Error(`All extension paths must start with a forward slash - "${subject.packageName}" used "${subject.item}".`)
    ^

Error: All extension paths must start with a forward slash - "govuk-frontend" used "[object Object]".
    at throwIfBadFilepath (/Users/ynda/code/github.com/dxw/dxw-prototype-lib/lib/extensions/extensions.js:75:11)
    at getFileSystemPath (/Users/ynda/code/github.com/dxw/dxw-prototype-lib/lib/extensions/extensions.js:145:3)
    at getPublicUrlAndFileSystemPath (/Users/ynda/code/github.com/dxw/dxw-prototype-lib/lib/extensions/extensions.js:152:19)
    at Array.map (<anonymous>)
    at Object.getPublicUrlAndFileSystemPaths (/Users/ynda/code/github.com/dxw/dxw-prototype-lib/lib/extensions/extensions.js:162:57)
    at setupPathsFor (/Users/ynda/code/github.com/dxw/dxw-prototype-lib/lib/middleware/extensions/extensions.js:7:14)
    at Object.<anonymous> (/Users/ynda/code/github.com/dxw/dxw-prototype-lib/lib/middleware/extensions/extensions.js:13:1)
    at Module._compile (node:internal/modules/cjs/loader:1469:14)
    at Module._extensions..js (node:internal/modules/cjs/loader:1548:10)
    at Module.load (node:internal/modules/cjs/loader:1288:32)
```

I was able to get the app running fine by adjusting some code in
lib/extensions/extensions.js and
lib/middleware/extensions/extensions.js, but the server logs showed some
missing file errors and this felt a bit hacky. These lib files were
generated when installing the app from govuk-prototype-kit@12.2.0 and
hadn't been touched since

Instead, this migrates the app to govuk-prototype-kit@latest (which
should be 13.16.2). It took a few attempts to get the migration to run
successfully, since some of the generated files had been edited. I
reverted some of these to their default to allow the migration to run
in full, then reapplied some of the changes we've made:
- Add custom `beforeContent` and `header` blocks in
  app/views/layout.html
- Add `serviceName` in app/config.json (previously in app/config.js)
- Add `module.exports = router` in app/routes.js
- Retain the Procfile but remove the build script since this no longer
  exists
- Retain the README.md but move it to the project root (and auto-format
  it)

This app has no tests, so it's likely that a Renovate upgrade PR was
incompatible with govuk-prototype-kit@12.2.0 and therefore broke things
quietly. This was only noticed when someone tried to look at the app on
Heroku