Set up Playwright
Breaking changes were merged to the app that caused it to be unable to
boot up. We didn't notice this until someone tried to load up the
deployed site. Non-major Renovate dependency PRs are merged
automatically on this repo, which is risky without at least minimal
automated testing. We had a CI workflow to test that the app built but
this evidently wasn't sufficient. This adds Playwright so that we can do
a basic 'homepage loads' kind of end-to-end test, and run this on CI for
pull requests and pushes to main