Remove build step from CI workflow
After the govuk-prototype-kit migration, this build command no longer
exists. This CI check has proven insufficient in any case: the app had
gotten into a buildable but non-runnable state. What we really need is a
check that the app can run, which will be added separately. This removes
this step/command to allow the workflow to continue passing in the
meantime